Refacto - moved identificationStart to Ubble folder
