+++ src/features/identityCheck/pages/identification/ubble/ComeBackLater.native.test.tsx
@@ -1,6 +1,6 @@
 import React from 'react'
 
-import { ComeBackLater } from 'features/identityCheck/pages/identification/identificationStart/ComeBackLater'
+import { ComeBackLater } from 'features/identityCheck/pages/identification/ubble/ComeBackLater'
 import { navigateToHomeConfig } from 'features/navigation/helpers'
 import { navigateFromRef } from 'features/navigation/navigationRef'
 import { amplitude } from 'libs/amplitude'
@@ -46,8 +46,8 @@ describe('ComeBackLater', () => {
   })
   it("should send a amplitude event when the 'M'identifier plus tard' button is pressed", async () => {
     const { getByText } = render(<ComeBackLater />)
-    // clear the amplitude log event mock to clear the first amplitude event screen_view_come_back_later
-    ;(amplitude.logEvent as jest.Mock).mockClear()
+      // clear the amplitude log event mock to clear the first amplitude event screen_view_come_back_later
+      ; (amplitude.logEvent as jest.Mock).mockClear()
 
     const button = getByText('M’identifier plus tard')
     fireEvent.press(button)
