+++ src/features/identityCheck/pages/identification/ubble/ExpiredOrLostID.native.test.tsx
@@ -1,6 +1,6 @@
 import React from 'react'
 
-import { ExpiredOrLostID } from 'features/identityCheck/pages/identification/identificationStart/ExpiredOrLostID'
+import { ExpiredOrLostID } from 'features/identityCheck/pages/identification/ubble/ExpiredOrLostID'
 import * as NavigationHelpers from 'features/navigation/helpers/openUrl'
 import { amplitude } from 'libs/amplitude'
 import { env } from 'libs/environment'
