+++ src/features/identityCheck/pages/identification/ubble/SelectIDStatus.native.test.tsx
@@ -1,7 +1,7 @@
 import React from 'react'
 
 import { navigate } from '__mocks__/@react-navigation/native'
-import { SelectIDStatus } from 'features/identityCheck/pages/identification/identificationStart/SelectIDStatus'
+import { SelectIDStatus } from 'features/identityCheck/pages/identification/ubble/SelectIDStatus'
 import { amplitude } from 'libs/amplitude'
 import { fireEvent, render, waitFor } from 'tests/utils'
