+++ src/features/identityCheck/pages/identification/ubble/SelectPhoneStatus.web.test.tsx
@@ -1,7 +1,7 @@
 import React from 'react'
 
 import { navigate } from '__mocks__/@react-navigation/native'
-import { SelectPhoneStatus } from 'features/identityCheck/pages/identification/identificationStart/SelectPhoneStatus.web'
+import { SelectPhoneStatus } from 'features/identityCheck/pages/identification/ubble/SelectPhoneStatus.web'
 import { checkAccessibilityFor, fireEvent, render, screen } from 'tests/utils/web'
 
 describe('SelectPhoneStatus', () => {
