+++ src/features/navigation/RootNavigator/subscriptionRoutes.ts
@@ -9,11 +9,11 @@ import { IdentityCheckEduConnectForm } from 'features/identityCheck/pages/identi
 import { IdentityCheckValidation } from 'features/identityCheck/pages/identification/educonnect/IdentityCheckValidation'
 import { withEduConnectErrorBoundary } from 'features/identityCheck/pages/identification/errors/eduConnect/EduConnectErrorBoundary'
 import { EduConnectErrors } from 'features/identityCheck/pages/identification/errors/eduConnect/EduConnectErrors'
-import { ComeBackLater } from 'features/identityCheck/pages/identification/identificationStart/ComeBackLater'
-import { ExpiredOrLostID } from 'features/identityCheck/pages/identification/identificationStart/ExpiredOrLostID'
-import { SelectIDOrigin } from 'features/identityCheck/pages/identification/identificationStart/SelectIDOrigin'
-import { SelectIDStatus } from 'features/identityCheck/pages/identification/identificationStart/SelectIDStatus'
-import { SelectPhoneStatus } from 'features/identityCheck/pages/identification/identificationStart/SelectPhoneStatus.web'
+import { ComeBackLater } from 'features/identityCheck/pages/identification/ubble/ComeBackLater'
+import { ExpiredOrLostID } from 'features/identityCheck/pages/identification/ubble/ExpiredOrLostID'
+import { SelectIDOrigin } from 'features/identityCheck/pages/identification/ubble/SelectIDOrigin'
+import { SelectIDStatus } from 'features/identityCheck/pages/identification/ubble/SelectIDStatus'
+import { SelectPhoneStatus } from 'features/identityCheck/pages/identification/ubble/SelectPhoneStatus.web'
 import { IdentityCheckUnavailable } from 'features/identityCheck/pages/identification/IdentityCheckUnavailable'
 import { IdentityCheckEnd } from 'features/identityCheck/pages/identification/ubble/IdentityCheckEnd'
 import { IdentityCheckPending } from 'features/identityCheck/pages/identification/ubble/IdentityCheckPending'
